refactor: move upgrade-only test helpers to upgrade_tests_helper.py
helpers.py had accumulated a lot of logic that is exclusively used by
the revision-pinned upgrade integration tests (the 5 run_*_scenario
orchestrators, force_leader_change, the plain-HTTP verification
functions, etc.), alongside genuinely shared infrastructure used by
non-upgrade tests too.

Move everything that has zero shared mutable state and zero non-upgrade
importers into a new tests/integration/upgrade_tests_helper.py:
assert_traefik_revision, verify_http_on_all_units, verify_http_on_unit,
leader_unit_name, force_leader_change, bring_up_traefik_without_certificate_provider,
and the 5 run_*_upgrade_scenario functions.

Left in helpers.py: the CA/certificate-signing machinery (generate_ca,
sign_csr, provide_certificate, provide_existing_certificate,
sign_csrs_and_provide_cert, get_outstanding_csrs) and the HTTPS
verification/bring-up functions (bring_up_certified_traefik,
bring_up_self_signed_traefik, verify_https_on_all_units,
verify_https_on_unit, pull_ssc_ca_certificate) since they share a
module-level CA-state global (ca_key/ca_cert/ca_cert_path/signed_certificates)
and two of them (verify_https_on_all_units, pull_ssc_ca_certificate) are
directly imported by non-upgrade tests (test_https_multi_unit.py,
test_tls.py) — splitting that cluster would have either broken the
global-state handoff or forced non-upgrade tests to import from a
module named upgrade_tests_helper.

Updated imports in all 19 consumer test files accordingly. Verified via
a pytest --collect-only diff against the pre-branch baseline: 103/103
tests, zero test node ID differences. Full tox -e lint pass clean.
