+++ tests/integration/test_upgrade_no_tls_from_280.py
@@ -18,7 +18,7 @@ Scenario:
 import jubilant
 import pytest
 from conftest import TRAEFIK_RESOURCES
-from helpers import run_no_tls_upgrade_scenario
+from upgrade_tests_helper import run_no_tls_upgrade_scenario
 
 SOURCE_REVISION = 280
