+++ tests/integration/test_upgrade_ssc_single_unit_from_280_via_298.py
@@ -12,10 +12,10 @@ from conftest import TRAEFIK_APP_NAME, TRAEFIK_RESOURCES
 from constants import MOCK_HOSTNAME, SOURCE_CHANNEL, TRAEFIK_CHARM
 from helpers import (
     all_settled,
-    assert_traefik_revision,
     bring_up_self_signed_traefik,
     verify_https_on_unit,
 )
+from upgrade_tests_helper import assert_traefik_revision
 
 logger = logging.getLogger(__name__)
